move thread/graph utils used by vox into thread and utils.graph
